fix(systemd/*): specify tag in `docker run` commands as well
While Deis systemd unit files specify the docker image tag in the
`docker pull` ExecPreStart command, they omit it in their `docker run`
command. This inadvertently pulls down *all* tagged images for the
docker container, wasting bandwidth. This is most evident when testing
on the v0.9.1 branch; on master it is effectively masked by the
semantics of the ":latest" tag.