Added provisioner for Deis on DigitalOcean
As DO does not support CoreOS we deploy CoreOS on top of Ubuntu 14.04 using kexec on boot to
switch to the CoreOS kernel and initrd. The state filesystem is a loop mounted btrfs.

The CoreOS bootstrapping is heavily based on Levi Aul's code:

  https://gist.github.com/tsutsu/490f35f48897df0f5173

We don't use Debian 7.0 as base distribution because kexec is unstable on the Debian
kernel.